state: split progress state into multiple files
