+++ state/progress/index.ts
@@ -1,0 +1,18 @@
+/**
+ * Public barrel for all progress state exports.
+ *
+ * This keeps external imports stable while internals stay split into focused
+ * modules.
+ */
+// Default values and enums.
+export * from './defaultProgress'
+// Shared lesson access helpers.
+export * from './helpers'
+// Base atoms and sync wiring.
+export * from './atoms'
+// Read-oriented derived atoms.
+export * from './selectors'
+// Write-oriented action atoms.
+export * from './actions'
+// Pure utility functions.
+export * from './utils'
